changing packed_data/value to int

+++ projects/cm_mcu/I2CCommunication.h
@@ -20,7 +20,7 @@
 int apollo_i2c_ctl_set_dev(uint8_t base);
 int apollo_i2c_ctl_r(uint8_t address, uint8_t nbytes, uint8_t data[4]);
 int apollo_i2c_ctl_reg_r(uint8_t address, uint8_t reg_address, uint8_t nbytes, uint8_t data[4]);
-int apollo_i2c_ctl_w(uint8_t address, uint8_t nbytes, uint8_t value);
-int apollo_i2c_ctl_reg_w(uint8_t address, uint8_t reg_address, uint8_t nbytes, uint8_t packed_data);
+int apollo_i2c_ctl_w(uint8_t address, uint8_t nbytes, int value);
+int apollo_i2c_ctl_reg_w(uint8_t address, uint8_t reg_address, uint8_t nbytes, int packed_data);
 
 #endif /* PROJECTS_CM_MCU_I2CCOMMUNICATION_H_ */
